show orginal build time in CLI when pressing help
